+++ App/Tests/BuildifyTests.swift
@@ -1,5 +1,5 @@
-@testable import xpm
 import XCTest
+@testable import xpm
 
 class xpmTests: XCTestCase {
 }