+++ App/xpmKit/Sources/Utils/FileHandler.swift
@@ -12,6 +12,13 @@ protocol FileHandling: AnyObject {
     /// - Returns: true if there's a file at the given path.
     func exists(_ path: AbsolutePath) -> Bool
 
+    /// Copies a file from one location to another.
+    ///
+    /// - Parameters:
+    ///   - from: the location the file is being copied from.
+    ///   - to: the location the file is being copied to.
+    func copy(from: AbsolutePath, to: AbsolutePath) throws
+
     /// Returns all the files using the glob pattern.
     ///
     /// - Parameters:
@@ -19,6 +26,17 @@ protocol FileHandling: AnyObject {
     ///   - glob: glob pattern.
     /// - Returns: list of paths that have been found matching the glob pattern.
     func glob(_ path: AbsolutePath, glob: String) -> [AbsolutePath]
+
+    /// Creates a folder at the given path.
+    ///
+    /// - Parameter path: path.
+    func createFolder(_ path: AbsolutePath) throws
+
+    /// Deletes the file at the given path.
+    ///
+    /// - Parameter path: path where the file is.
+    /// - Throws: an error if the deletion fails.
+    func delete(_ path: AbsolutePath) throws
 }
 
 /// Default file handler implementing FileHandling.
@@ -36,6 +54,15 @@ final class FileHandler: FileHandling {
         return FileManager.default.fileExists(atPath: path.asString)
     }
 
+    /// Copies a file from one location to another.
+    ///
+    /// - Parameters:
+    ///   - from: the location the file is being copied from.
+    ///   - to: the location the file is being copied to.
+    func copy(from: AbsolutePath, to: AbsolutePath) throws {
+        try FileManager.default.copyItem(atPath: from.asString, toPath: to.asString)
+    }
+
     /// Returns all the files using the glob pattern.
     ///
     /// - Parameters:
@@ -44,5 +71,22 @@ final class FileHandler: FileHandling {
     /// - Returns: list of paths that have been found matching the glob pattern.
     func glob(_ path: AbsolutePath, glob: String) -> [AbsolutePath] {
         return path.glob(glob)
+    }
+
+    /// Creates a folder at the given path.
+    ///
+    /// - Parameter path: path.
+    func createFolder(_ path: AbsolutePath) throws {
+        try FileManager.default.createDirectory(atPath: path.asString,
+                                                withIntermediateDirectories: true,
+                                                attributes: nil)
+    }
+
+    /// Deletes the file at the given path.
+    ///
+    /// - Parameter path: path where the file is.
+    /// - Throws: an error if the deletion fails.
+    func delete(_ path: AbsolutePath) throws {
+        try FileManager.default.removeItem(atPath: path.asString)
     }
 }