+++ App/xpmKit/Tests/Fixtures/xpm.framework/Headers/xpm.h
@@ -1,0 +1,19 @@
+//
+//  xpm.h
+//  xpm
+//
+//  Created by Pedro Piñera Buendía on 03.06.18.
+//  Copyright © 2018 com.xcodepm. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for xpm.
+FOUNDATION_EXPORT double xpmVersionNumber;
+
+//! Project version string for xpm.
+FOUNDATION_EXPORT const unsigned char xpmVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <xpm/PublicHeader.h>
+
+