+++ App/xpmKit/Tests/Loader/GraphLoadingErrorTests.swift
@@ -1,7 +1,7 @@
 import Basic
 import Foundation
-@testable import xpmKit
 import XCTest
+@testable import xpmKit
 
 final class GraphLoadingErrorTests: XCTestCase {
     func test_description_returns_the_right_value_when_manifestNotFound() {