+++ App/xpmKit/Tests/Utils/Mocks/MockRersourceLocator.swift
@@ -6,6 +6,8 @@ final class MockResourceLocator: ResourceLocating {
     var projectDescriptionStub: (() throws -> AbsolutePath)?
     var cliPathCount: UInt = 0
     var cliPathStub: (() throws -> AbsolutePath)?
+    var embedPathCount: UInt = 0
+    var embedPathStub: (() throws -> AbsolutePath)?
     var appPathCount: UInt = 0
     var appPathStub: (() throws -> AbsolutePath)?
 
@@ -17,6 +19,11 @@ final class MockResourceLocator: ResourceLocating {
     func cliPath() throws -> AbsolutePath {
         cliPathCount += 1
         return try cliPathStub?() ?? AbsolutePath("/")
+    }
+
+    func embedPath() throws -> AbsolutePath {
+        embedPathCount += 1
+        return try embedPathStub?() ?? AbsolutePath("/")
     }
 
     func appPath() throws -> AbsolutePath {