+++ App/xpmKit/Tests/Validator/ProjectValidatorTests.swift
@@ -1,7 +1,7 @@
 import Basic
 import Foundation
-@testable import xpmKit
 import XCTest
+@testable import xpmKit
 
 final class ProjectValidationErrorTests: XCTestCase {
     func test_description_whenDuplicatedTargets() {